initial commit for home page
